Fix %d format specifiers to %zu for size_t variables in log_e calls
Agent-Logs-Url: https://github.com/matthias-bs/BresserWeatherSensorReceiver/sessions/78e1cfaa-aa95-4e7a-9933-73287b1a188f

+++ examples/BresserWeatherSensorMQTT/src/mqtt_comm.cpp
@@ -380,17 +380,17 @@ void publishWeatherdata(bool complete, bool retain)
         }
         size_t json_size = serializeJson(jsonSensor, payloadSensor, sizeof(payloadSensor));
         if (json_size < 5) {
-            log_e("Failed to serialize JSON: size = %d", json_size);
+            log_e("Failed to serialize JSON: size = %zu", json_size);
         }
         size_t extra_size = serializeJson(jsonExtra, payloadExtra, sizeof(payloadExtra));
 
         if (json_size >= PAYLOAD_SIZE - 1)
         {
-            log_e("payloadSensor (%d) >= PAYLOAD_SIZE (%d). Payload truncated!", json_size, PAYLOAD_SIZE);
+            log_e("payloadSensor (%zu) >= PAYLOAD_SIZE (%zu). Payload truncated!", json_size, (size_t)PAYLOAD_SIZE);
         }
         if (extra_size >= PAYLOAD_EXTRA_SIZE - 1)
         {
-            log_e("payloadExtra (%d) >= PAYLOAD_EXTRA_SIZE (%d). Payload truncated!", extra_size, PAYLOAD_EXTRA_SIZE);
+            log_e("payloadExtra (%zu) >= PAYLOAD_EXTRA_SIZE (%zu). Payload truncated!", extra_size, (size_t)PAYLOAD_EXTRA_SIZE);
         }
 
         // Try to map sensor ID to name to make MQTT topic explanatory
@@ -427,7 +427,7 @@ void publishWeatherdata(bool complete, bool retain)
     size_t combined_size = serializeJson(jsonCombined, payloadCombined, sizeof(payloadCombined));
     if (combined_size >= PAYLOAD_SIZE - 1)
     {
-        log_e("payloadCombined (%d) >= PAYLOAD_SIZE (%d). Payload truncated!", combined_size, PAYLOAD_SIZE);
+        log_e("payloadCombined (%zu) >= PAYLOAD_SIZE (%zu). Payload truncated!", combined_size, (size_t)PAYLOAD_SIZE);
     }
     snprintf(mqtt_topic, sizeof(mqtt_topic), "%s/%s",
              Hostname.c_str(), mqttTopics.pubCombined);
